Store only 32 bit time for events (work out full 64 bits in event loop)             Increase event buffer size to 128 (from 64)

+++ src/jsdevices.c
@@ -205,7 +205,7 @@ void jshPushIOEvent(IOEventFlags channel, JsSysTime time) {
     return; // queue full - dump this event!
   }
   ioBuffer[ioHead].flags = channel;
-  ioBuffer[ioHead].data.time = time;
+  ioBuffer[ioHead].data.time = (unsigned int)time;
   ioHead = nextHead;
 }
